MVC pattern observed
