+++ app.js
@@ -1,5 +1,6 @@
 const express = require('express');
-const { getTopics, getArticleById, getArticleCommentsById, getArticles, getUsers, patchArticleById } = require('./controllers/controllers');
+const { getTopics, getUsers, } = require('./controllers/controllers');
+const { getArticleById, getArticleCommentsById, getArticles, patchArticleById } = require('./controllers/articles-controller');
 
 
 const app = express();
@@ -7,14 +8,14 @@ const app = express();
 // middleware
 app.use(express.json());
 
-//get requests
+// GET requests
 app.get('/api/topics', getTopics);
 app.get('/api/articles/:article_id', getArticleById);
 app.get('/api/articles/:article_id/comments', getArticleCommentsById);
 app.get('/api/users', getUsers);
 app.get('/api/articles', getArticles)
 
-// patch requests
+// PATCH requests
 app.patch('/api/articles/:article_id', patchArticleById);
 
 app.use((err, req, res, next) => {
@@ -25,7 +26,6 @@ app.use((err, req, res, next) => {
     next(err);
   };
 });
-
 
 // handles unexpected errors - check test log for error codes
 app.use((err, req, res, next) => {
