+++ controllers/controllers.js
@@ -1,26 +1,11 @@
 const { articleData } = require('../db/data/test-data');
 const articles = require('../db/data/test-data/articles');
 const topics = require('../db/data/test-data/topics');
-const { selectTopics, selectArticles, selectArticleById, selectArticleCommentsById, selectUsers, updateArticleById } = require('../models/models');
+const { selectTopics, selectUsers, } = require('../models/models');
 
 exports.getTopics = (req, res, next) => {
   selectTopics().then((topics) => {
     res.status(200).send(topics)
-  });
-};
-
-exports.getArticleById = (req, res, next) => {
-  const { article_id } = req.params;
-  selectArticleById(article_id).then((article) => {
-      res.status(200).send(article.rows[0]);
-  });
-};
-
-exports.getArticleCommentsById = (req, res, next) => {
-  const { article_id } = req.params;
-  selectArticleCommentsById(article_id).then((article) => {
-    console.log(article.rows);
-      res.status(200).send(article.rows);
   });
 };
 
@@ -30,16 +15,5 @@ exports.getUsers = (req, res, next) => {
   });
 };
 
-exports.getArticles = (req, res, next) => {
-  selectArticles().then((articles) => {
-    res.status(200).send(articles)
-  })
-}
 
-exports.patchArticleById = (req, res, next) => {
-  const { inc_votes: votes } = req.body;
-  const { article_id } = req.params;
-  updateArticleById(article_id, votes).then((article) => {
-    res.status(201).send(article.rows[0])
-  });
-}
+
